chore: StepCircle 분리 (#11)
