+++ BBANGZIP/BBANGZIP/Sources/Presentation/ProgressBar/StepCircle.swift
@@ -1,0 +1,27 @@
+//
+//  StepCircle.swift
+//  BBANGZIP
+//
+//  Created by 최유빈 on 1/14/25.
+//  Copyright © 2025 com.bbangzip. All rights reserved.
+//
+
+import SwiftUI
+
+struct StepCircle: View {
+    var step: Int
+    var complete: Bool = true
+    
+    var body: some View {
+        Circle()
+            .overlay {
+                BbangText(
+                    "\(step)",
+                    fontType: .caption2,
+                    color: complete ? Color(BBANGZIPAsset.Assets.staticWhite.color) : Color(BBANGZIPAsset.Assets.labelDisable.color)
+                )
+            }
+            .foregroundStyle(complete ? Color(BBANGZIPAsset.Assets.statusPositive.color) : Color(BBANGZIPAsset.Assets.fillAlternative.color))
+            .frame(width: 20, height: 20)
+    }
+}
